requirements.txt
Django==4.1.2
django-environ==0.9.0
django-picklefield==3.1
djangorestframework==3.14.0
requests==2.28.1
